+++ src/packages/core/workspace/components/workspace-action/common/index.ts
@@ -1,1 +1,2 @@
 export * from './save/index.js';
+export * from './workspace-action-base.js';
