+++ src/packages/core/workspace/components/workspace-action/common/workspace-action-base.ts
@@ -1,5 +1,5 @@
-import type { UmbWorkspaceContextInterface } from '../../../../workspace-context/index.js';
-import { UMB_WORKSPACE_CONTEXT } from '../../../../workspace-context/index.js';
+import type { UmbWorkspaceContextInterface } from '../../../workspace-context/index.js';
+import { UMB_WORKSPACE_CONTEXT } from '../../../workspace-context/index.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbBaseController } from '@umbraco-cms/backoffice/class-api';
 import type { UmbApi } from '@umbraco-cms/backoffice/extension-api';
