+++ src/packages/core/workspace/components/workspace-action/shared/workspace-action/index.ts
@@ -1,2 +1,1 @@
-export * from './workspace-action-base.js';
 export * from './workspace-action.element.js';
